Wartung und Alarm spiegeln ihren Freitext ins Gedächtnis
Der Grund eines Degradationsverlaufs lebt ausschliesslich im Freitext, nie als
Datenpunkt (§12.5, Beobachtungsgrenze). Ein Sensor sieht nicht, dass ein
Ersatzfett mit zu niedriger Grundölviskosität eingefüllt wurde — das steht in der
Beschreibung oder nirgends. Wartung und Alarm geben ihre Beschreibung bzw.
Meldung deshalb an das Gedächtnis weiter, so wie die Schichtnotiz seit dem
24.08.2026 ihren Text weitergibt. Damit kann das Gedächtnis die Ursache eines
Vorgangs kennen, und "hatten wir das schon mal" hat etwas zu finden.

Belegt statt vermutet: Mit reiner Struktur brachte das Gedächtnis als vierte
Archiv-Quelle auf keiner von 18 Goldset-Anfragen einen zusätzlichen zutreffenden
Treffer (Register C-050). Produktionslauf und Drift bleiben bei der Struktur, weil
ihre Quelle keinen Freitext führt; Ereigniskette und Werker-Empfehlung bleiben es
aus einem anderen Grund — ihr Text wäre eine Ableitung des Systems, und sie zu
spiegeln hiesse, das Gedächtnis mit den eigenen Schlüssen zu speisen und diese
später als Beleg wiederzufinden. Ein Test hält die Begrenzung fest.

Maskiert wie der Notiz-Freitext: Beide Felder laufen vor der Spiegelung durch den
Redactor. Was das System verlässt, wird gleich behandelt — der Spiegel darf nicht
die schwächere Grenze sein. Die Quellzeile bleibt unverändert; §8 nannte diesen
Fall bereits ("bei Bedarf Redactor später auf diese Felder ausweiten"). Belegt ist
die Verdrahtung, nicht die Erkennungsgüte — die hängt am Marker ner, den der
Pflichtlauf deselektiert, und wird nirgends behauptet.

Der Feldzugriff ist weich, als einzige Stelle in content.py: Altzeilen tragen die
Felder nicht, und ein harter Zugriff liesse den gesamten Altbestand aus der
Spiegelung fallen, um ein Feld zu gewinnen, das er ohnehin nicht hat. Ohne
Freitext entsteht wortgleich der bisherige Satz — eingefordert samt Gegenprobe,
denn schon ein angehängtes Leerzeichen legte einen zweiten Eintrag für dasselbe
Ereignis an.

Neu ist ausserdem `python -m foreman.substrate.nachtrag` für den Altbestand. Der
Backfill kann ihn nicht erreichen: Er wählt substrate_ref IS NULL und baut aus der
gespeicherten Nutzlast. Der Nachtrag liest den Freitext aus der Quelltabelle,
löscht die alte Erinnerung und gibt die Zeile zur Neuspiegelung frei — in dieser
Reihenfolge, damit ein Abbruch dazwischen wiederholbar bleibt statt eine verwaiste
Erinnerung zu hinterlassen. Eine gestörte Gegenstelle verbraucht keine Zeile; ein
Test prüft das an der Auswahl des nächsten Laufs, nicht am Zähler.

24 Tests, davon 5 Gegenproben mit Kontrollpunkt. Volle Suite 1107 grün bei 93,98 %.

+++ tools/archiv_guete/baue_goldset.py
@@ -25,9 +25,7 @@ def main() -> None:
     bewerter_liste = roh["result"]["urteile"] if "result" in roh else roh["urteile"]
 
     # stimmen[anfrage_id][schluessel] = [(bewerter, stufe), ...]
-    stimmen: dict[str, dict[str, list[tuple[str, int]]]] = defaultdict(
-        lambda: defaultdict(list)
-    )
+    stimmen: dict[str, dict[str, list[tuple[str, int]]]] = defaultdict(lambda: defaultdict(list))
     namen = []
     for eintrag in bewerter_liste:
         name = eintrag["bewerter"]
@@ -44,7 +42,11 @@ def main() -> None:
         goldset[aid] = {}
         for schluessel, liste in sorted(stimmen[aid].items()):
             if len(liste) >= MINDESTSTIMMEN:
-                stufe = int(round(statistics.median([s for _, s in liste])))
+                # Bei Uneinigkeit (Stufen 1 und 2) rundet round() auf 2 — im
+                # Zweifel die hoehere Stufe. Das ist eine Entscheidung, keine
+                # Nebenwirkung: Zwei Beurteiler, die einen Eintrag ueberhaupt
+                # nennen, halten ihn fuer erheblich.
+                stufe = round(statistics.median([s for _, s in liste]))
                 goldset[aid][schluessel] = stufe
                 if len(liste) == MINDESTSTIMMEN:
                     knapp.append(f"{aid}/{schluessel}")
@@ -66,18 +68,14 @@ def main() -> None:
     for aid, treffer in goldset.items():
         gesamt += len(treffer)
         zwei = sum(1 for s in treffer.values() if s == 2)
-        print(
-            f"{aid:6s} {len(treffer):4d} {zwei:7d}  {', '.join(sorted(treffer))[:52]}"
-        )
+        print(f"{aid:6s} {len(treffer):4d} {zwei:7d}  {', '.join(sorted(treffer))[:52]}")
     print("-" * 76)
     print(f"Relevante Zuordnungen gesamt : {gesamt}")
     print(f"davon knapp (genau {MINDESTSTIMMEN} Stimmen) : {len(knapp)}")
     print(f"verworfen (nur 1 Stimme)     : {len(verworfen)}")
 
     # Uebereinstimmung: Anteil der Zuordnungen, die ALLE Bewerter nannten.
-    alle = sum(
-        1 for a in stimmen.values() for liste in a.values() if len(liste) == len(namen)
-    )
+    alle = sum(1 for a in stimmen.values() for liste in a.values() if len(liste) == len(namen))
     genannt = sum(len(a) for a in stimmen.values())
     print(
         f"\nUebereinstimmung: {alle} von {genannt} genannten Zuordnungen von ALLEN {len(namen)} "
